requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto3==1.7.19
botocore==1.10.19
celery==3.1.26.post2
Django==2.0.5
django-celery==3.2.2
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
kombu==3.0.37
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.3
pytz==2018.4
s3transfer==0.1.13
six==1.11.0
